+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Generation based on all class in a package (packageName parameter)
+
 ## [1.3.2] - 2024-04-01
 
 ### Added